REF: extract RsDefaultValueBuilder
